added test qseis store vs qseis2d store, remove old test
